+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprDateFromUnix.java
@@ -6,8 +6,6 @@ import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
 import io.github.syst3ms.skriptparser.util.SkriptDate;
 
-import java.math.BigInteger;
-
 /**
  * The date from a (unix) timestamp.
  * The default timestamp returns the amount of <b>milliseconds</b> since the Unix Epoch.
@@ -30,14 +28,14 @@ public class ExprDateFromUnix implements Expression<SkriptDate> {
 		);
 	}
 
-	private Expression<BigInteger> timestamp;
+	private Expression<Integer> timestamp;
 	private boolean unix;
 
 	@SuppressWarnings("unchecked")
 	@Override
 	public boolean init(Expression<?>[] expressions, int matchedPattern, ParseContext parseContext) {
 		unix = parseContext.getNumericMark() == 1;
-		timestamp = (Expression<BigInteger>) expressions[0];
+		timestamp = (Expression<Integer>) expressions[0];
 		return true;
 	}
